COOPR-518: minor PR fixes
